Комментарии переехали в смарт-компонент
